requirements.txt
torch>=2.0.0
transformers>=4.30.0
kobert-transformers
gluonnlp
pandas
numpy
scikit-learn
tqdm
pillow
requests
spotipy
python-dotenv
openai>=1.0.0
streamlit